Tweak probing logic
This changes probing logic so that we send multiple packets (2 for now)
in every packet number space on PTO... But only the first forces a
probe.

This should avoid the problem we had regarding retransmissions that
didn't have space for their payload in a single packet on PTO.  That's
something we relied on for the reorder_handshake test.

This change produced some pretty significant instability though.  I see
anywhere between 4 and 8 tests failing as a result.  Errors are not all
reproducible.  That might be because I added 100 to the starting packet
number in the client Initial, who knows.  Here's the list:

    connection::tests::ackrate::ack_rate_persistent_congestion
    connection::tests::cc::cc_cong_avoidance_recovery_period_unchanged
    connection::tests::cc::single_packet_on_recovery
    connection::tests::handshake::reorder_05rtt_with_0rtt
    connection::tests::handshake::reorder_handshake
    connection::tests::recovery::loss_time_past_largest_acked
    connection::tests::recovery::pto_handshake_complete
    connection::tests::resumption::remember_smoothed_rtt